do not revert using do not revert useSuspenseQuery in FlagAdminList
